+++ member-service/member-service-client/member-service-client-api/src/main/java/com/project/imdang/member/domain/client/MemberData.java
@@ -19,7 +19,7 @@ public class MemberData {
     private String birthDate;
     private String gender;
     private String deviceToken;
-    private int accusedCount;
-    private int insightCount;
+    private Long accusedCount;
+    private Long insightCount;
     private LocalDate latestInsightCreateDate;
 }
